backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "medlens-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.1",

    # GraphQL
    "strawberry-graphql[fastapi]>=0.225.0",

    # Database (PostgreSQL on AWS RDS)
    "sqlalchemy>=2.0.28",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",

    # Validation
    "pydantic>=2.6.3",
    "pydantic-settings>=2.2.1",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",

    # AWS (S3 image storage, EventBridge, future Lambda/ECS)
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
